Split network tests into one file per source module
braincell/network/ has seven source modules and, until now, zero test files
whose name matched any of them: all 57 tests lived in runtime_test.py and
topology_test.py. Refile them by the module each test calls directly.

  core_test.py        Population and Connection validation, reprs
  lowering_test.py    LoweringTest wholesale (validation, delay quantization)
  edges_test.py       EdgeSet construction/exports plus Network.add_edges
  projections_test.py Projection, ContactTable, contact methods
  engine_test.py      Network state, run caches, delivery timing, backends
  delivery_test.py    the two behaviours delivery.py exposes directly

delivery_test.py is deliberately thin. Only two of the 57 tests call
delivery.population_spike or delivery.resolve_event_backend directly; the
ring-buffer arrival machinery is covered only end-to-end through Network.run.
Filing them honestly makes that gap visible instead of hiding it inside a
catch-all runtime_test.py.

The eight cell/tree/solver builders shared across these files move to the new
network/_testing.py, deduplicating runtime_test's _build_tree against
topology_test's _build_cell.

Collected test count is unchanged at 57 for this package.
